Bugfix: Ignore the 'default_branch' option
When specifying a default branch on a repository which is still empty, the execution
fails with the following error message:
'Cannot update default branch for an empty repository. Please init the repository and push first'

This change avoids that users have to debug this problem and/or create new repositories in
three steps (1. ‘create repo PR’, 2. add branch , 3. ‘set default branch’ PR).

Secondly this change adds a bit more detailed error-debug-output to simplify
debugging situations.

Signed-off-by: Marc Schöchlin <schoechlin@osb-alliance.com>